- Btrfs error handling update:   - Quilt-refresh patchset   - Added simplification for ->submit_bio_hook   - Added missing btree locking panic for convert_extent_bit.   - Fixed typos in Subject headers
suse-commit: d52421aa8c52bd1fbbc9e9dd92428e2b1a5f759b

+++ fs/btrfs/ordered-data.c
@@ -64,7 +64,7 @@ NORET_TYPE static void ordered_data_tree_panic(struct inode *inode, int errno,
 {
 	struct btrfs_fs_info *fs_info = btrfs_sb(inode->i_sb)->fs_info;
 	btrfs_panic(fs_info, errno, "Inconsistency in ordered tree at offset "
-		    "%llu\n", offset);
+		    "%llu\n", (unsigned long long)offset);
 }
 
 /*
